convert python code to 'classify' package, add first test
