+++ src/classify/__init__.py
@@ -1,0 +1,1 @@
+from .classifuncs import sigm, getxy, gradient_descent
